Refactor CLI structure and add SUR command
Moved main CLI logic from CLI.swift to a new SUR.swift file, introducing the SUR command for finding unused resources. Renamed RewriteCLI.swift to RToXcode.swift and updated references to reflect the new command structure. CLI.swift now delegates to SUR and RToXcode as subcommands, improving modularity and clarity.
